+++ functions/AVG_list.py
